CSV convertation has been fixed

+++ services/json_convertor.py
@@ -8,11 +8,18 @@ def convert(url: str) -> None:
     file_exists = os.path.exists('./downloads/events.csv')
     json_path = ai_response(url)
     df = pd.read_json(json_path)
-    df.to_csv(
-        './downloads/events.csv',
-          encoding='utf-8',
-         index=False,
-         mode='a',
-         header=not file_exists)
+    if not file_exists:
+        df.to_csv(
+            './downloads/events.csv',
+            index=False,
+            encoding='utf-8'
+        )
+    else:
+        df.to_csv(
+            './downloads/events.csv',
+            encoding='utf-8',
+            index=False,
+            mode='a',
+            header=not file_exists)
     remove_file()
-    print(json_path)
+
